# api/openapi.yaml
openapi: 3.1.0
info:
  title: Deck Fleet Orchestrator
  version: 0.1.0
  summary: HTTP API for the deck-fleet orchestrator and per-deck executor.
  description: |
    Machine-readable contract for the deck-fleet HTTP surface. The canonical
    prose is in `docs/API.md`; this file is the source of truth for generated
    Go and TypeScript types.

    Two boundaries, two API styles:

    - **Operator-facing API** (`/api/*`): RESTful resource shapes for reads,
      RPC-style action endpoints for state transitions. Clients poll for
      live updates (cadence in `docs/API.md`).
      Consumed by the React UI and by `curl`-driven demos.
    - **Executor-facing API** (`/executor/*`): the orchestrator hosts
      `/executor/poll`, `/executor/events`, `/executor/heartbeat`. Each
      executor process hosts `/executor/state` and `/executor/abort/{attempt_id}`
      that the orchestrator dials on suspicion / cancel.

    See `DESIGN.md` for the asymmetric pull/push topology and the state-machine
    transitions every operator action must respect.
  contact:
    name: deck-fleet
    url: https://example.invalid/deck-fleet
servers:
  - url: http://localhost:8080
    description: Orchestrator (default for /api/* and /executor/poll, /events, /heartbeat)
  - url: http://deck-1.lab.internal:8001
    description: Executor (per-deck; for /executor/state and /executor/abort)
...
paths:
  /api/runs:
    post:
      tags: [runs]
      operationId: submitRun
      summary: Submit a DAG.
      description: |
        Accept a new DAG, validate it, persist it, set initial state.
        Idempotent on the submission `id`: re-submitting the same id returns
        `409 DUPLICATE_RESOURCE` with the existing run in `details.current_state`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DagSubmission'
      responses:
        '201':
          description: Run created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Run'
        '400':
          description: Malformed or schema-invalid request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Duplicate id.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '413':
          description: Request body exceeds 1 MB.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: DAG semantic validation failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalErrorResponse'
        '503':
          $ref: '#/components/responses/DegradedModeResponse'
    get:
      tags: [runs]
      operationId: listRuns
      summary: List runs.
      parameters:
        - name: status
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/RunStatus'
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
      responses:
        '200':
          description: Run list.
          content:
            application/json:
              schema:
                type: object
                required: [runs]
                properties:
                  runs:
                    type: array
                    items:
                      $ref: '#/components/schemas/RunSummary'
        '400':
          description: Bad query parameter.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalErrorResponse'
...
  /api/decks:
    get:
      tags: [decks]
      operationId: listDecks
      summary: Fleet view of decks with current health and current job.
      responses:
        '200':
          description: Decks list.
          content:
            application/json:
              schema:
                type: object
                required: [decks]
                properties:
                  decks:
                    type: array
                    items:
                      $ref: '#/components/schemas/Deck'
        '500':
          $ref: '#/components/responses/InternalErrorResponse'
  /api/decks/{deck_id}/release:
    post:
      tags: [decks]
      operationId: releaseDeck
      summary: Release a slot back to EMPTY.
      description: |
        Operator-deliberate vacate: clears `endpoint_url` and
        `last_heartbeat_at`, sets `last_known_health = EMPTY`, emits
        `DECK_HEALTH_CHANGED` if the prior state was non-EMPTY.
        Idempotent on already-EMPTY slots. Refuses with
        `409 SLOT_HAS_INFLIGHT_WORK` when any deck_job in
        `DISPATCHED` / `RUNNING` / `AMBIGUOUS` references the slot --
        the operator must resolve in-flight work first so released
        attempts aren't silently abandoned.

        Wired by the supervisor's detach path so a deliberate "remove
        executor" round-trips to EMPTY instead of aging through
        STALE -> UNREACHABLE. UNREACHABLE then unambiguously means
        "executor crashed without operator action," restoring the
        rubric's commitment to operator-visible state.
      parameters:
        - $ref: '#/components/parameters/DeckId'
      responses:
        '200':
          description: Slot released (or already EMPTY).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Deck'
        '404':
          description: Unknown deck_id.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Slot has in-flight work; resolve before release.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '410':
          description: Slot is decommissioned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalErrorResponse'
...
  /api/state:
    get:
      tags: [live]
      operationId: getState
      summary: Bootstrap or delta-poll the global console state.
      description: |
        Returns a full snapshot (`runs`, `decks`, recent `events`,
        `server_seq`) when `since_seq` is 0 or absent. Returns only
        events with `seq > since_seq` otherwise. The client uses
        `server_seq` from each response as the next request's
        `since_seq`. Gap detection (event seq != last+1) and unknown
        event kinds prompt a re-bootstrap by resetting `since_seq` to 0.

        Drives the live React console: a single 1-second poll replaces
        the per-resource refetch hooks the early prototype used.
      parameters:
        - name: since_seq
          in: query
          required: false
          schema:
            type: integer
            format: int64
            minimum: 0
            default: 0
        - name: limit
          in: query
          required: false
          description: |
            Maximum number of events to return in this response.
            Caps to 1000. Clients that see `events.length == limit`
            should poll again immediately (or re-bootstrap if the
            event log has rotated).
          schema:
            type: integer
            minimum: 1
            maximum: 1000
            default: 500
      responses:
        '200':
          description: State snapshot or delta.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StateSnapshot'
        '500':
          $ref: '#/components/responses/InternalErrorResponse'
  /api/runs/{id}/state:
    get:
      tags: [live]
      operationId: getRunState
      summary: Bootstrap or delta-poll a run-detail screen.
      description: |
        Same pattern as `GET /api/state` but scoped to one run: the
        snapshot carries the full `Run` (with recent attempts) and
        only the events that touch that run. Used by the run-detail
        screen so the global fleet payload is not re-sent on every
        tick.
      parameters:
        - $ref: '#/components/parameters/RunId'
        - name: since_seq
          in: query
          required: false
          schema:
            type: integer
            format: int64
            minimum: 0
            default: 0
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 1000
            default: 500
      responses:
        '200':
          description: Run state snapshot or delta.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunStateSnapshot'
        '404':
          description: Run not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalErrorResponse'
components:
  parameters:
    RunId:
      name: id
      in: path
      required: true
      schema:
        type: string
...
    DeckId:
      name: deck_id
      in: path
      required: true
      schema:
        type: string
  responses:
    InternalErrorResponse:
      description: Unhandled server error.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    DegradedModeResponse:
      description: Orchestrator in degraded mode; mutations refused.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    RunStatus:
      type: string
      description: |
        Materialized run status. Terminal statuses are `COMPLETED` (all
        deck_jobs succeeded) and `CANCELLED` (operator cut the run).

        `FAILED` and `AMBIGUOUS` are non-terminal: the run is awaiting an
        operator decision (retry / resolve, or cancel). `terminal_at` is
        only stamped on terminal statuses; for FAILED / AMBIGUOUS it is
        null until the operator either resumes the run (retry / resolve)
        or terminalizes it via cancel. See DESIGN.md.
      enum:
        - PENDING
        - RUNNING
        - COMPLETED
        - FAILED
        - AMBIGUOUS
        - CANCELLED
...
    DeckJob:
      type: object
      description: A deck_job row plus recent attempts. See `DESIGN.md`.
      required:
        - id
        - deck_id
        - depends_on
        - steps
        - status
        - version
      additionalProperties: false
      properties:
        id:
          type: string
        deck_id:
          type: string
        depends_on:
          type: array
          items:
            type: string
        steps:
          type: array
          items:
            $ref: '#/components/schemas/Step'
        status:
          $ref: '#/components/schemas/DeckJobStatus'
        current_attempt_id:
          type: string
          format: uuid
          nullable: true
        version:
          type: integer
          format: int64
        error:
          type: string
          nullable: true
        last_completed_step:
          type: integer
          format: int64
          description: |
            Bootstrap denormalization of the most recent committed step cursor.
            0 means no steps have completed (job not yet running or just dispatched).
            Updated by the orchestrator each time it receives a STEP_COMPLETED
            executor event; also set to `total_steps` when a COMPLETED event lands
            so the UI never shows a partial count on a finished job.
        total_steps:
          type: integer
          format: int64
          description: |
            Total number of steps in this deck_job. Set at submit time from
            `len(deck_job.steps)`. Combined with `last_completed_step` this
            allows the UI to render "step 2/5" without replaying the event log.
        ambiguous_reason:
          type: string
          enum:
            - DEADLINE_ELAPSED
            - EXECUTOR_REPORTED_UNKNOWN
            - DEADLINE_EXCEEDED
          nullable: true
          description: |
            Set when status is AMBIGUOUS, identifying which liveness path
            declared the attempt unreachable: DEADLINE_EXCEEDED (the
            orchestrator's per-attempt hang ceiling fired and the executor
            still reported IN_PROGRESS), DEADLINE_ELAPSED (the deck went
            UNREACHABLE and the AmbiguousDeadline expired with no
            authoritative outcome), or EXECUTOR_REPORTED_UNKNOWN (the
            executor explicitly said it lost track of the attempt).
            Cleared (null) when the operator resolves or retries.
        recent_attempts:
          type: array
          description: Newest first. Only present on `GET /api/runs/{id}` responses.
          items:
            $ref: '#/components/schemas/JobAttempt'
...
    CurrentJob:
      type: object
      description: The deck_job currently occupying a deck's slot. Null when the deck is idle.
      required: [run_id, job_id, status]
      additionalProperties: false
      properties:
        run_id:
          type: string
        job_id:
          type: string
        status:
          $ref: '#/components/schemas/DeckJobStatus'
    Deck:
      type: object
      description: |
        Fleet view row. See `DESIGN.md`. With the
        decks-as-slots model, `endpoint_url` and `last_heartbeat_at` are
        nullable: a slot that has never been attached (or whose executor
        has been detached long enough for the liveness sweep to clear
        its scalars) has neither, and reports `last_known_health = EMPTY`.
        `decommissioned_at` is non-null only for retired slots; the
        operator-facing fleet view filters them out by default.
      required:
        - id
        - first_seen_at
        - last_known_health
      additionalProperties: false
      properties:
        id:
          type: string
        first_seen_at:
          type: string
          format: date-time
        last_heartbeat_at:
          type: string
          format: date-time
          nullable: true
        last_known_health:
          $ref: '#/components/schemas/DeckHealth'
        last_claimed_attempt_id:
          type: string
          format: uuid
          nullable: true
        endpoint_url:
          type: string
          nullable: true
          description: |
            The advertise URL the orchestrator dials for /executor/state
            and /executor/abort. Populated on first heartbeat; nullable
            until then (empty slot) and cleared by long-detach aging.
        decommissioned_at:
          type: string
          format: date-time
          nullable: true
          description: |
            Operator-set retirement timestamp. Decommissioned slots are
            hidden from the default fleet listing and reject heartbeats
            with 410 Gone.
        current_job:
          allOf:
            - $ref: '#/components/schemas/CurrentJob'
          nullable: true

    # --------------------------------------------------------------------
    # DAG submission
    # --------------------------------------------------------------------
    DagJobSubmission:
      type: object
      description: A deck_job as submitted by the operator (no server-added fields).
      required: [id, deck_id, depends_on, steps]
      additionalProperties: false
      properties:
        id:
          type: string
        deck_id:
          type: string
        depends_on:
          type: array
          items:
            type: string
        steps:
          type: array
          items:
            $ref: '#/components/schemas/Step'
    DagSubmission:
      type: object
      description: The DAG payload an operator submits to `POST /api/runs`.
      required: [id, deck_jobs]
      additionalProperties: false
      properties:
        id:
          type: string
        deck_jobs:
          type: array
          items:
            $ref: '#/components/schemas/DagJobSubmission'

    # --------------------------------------------------------------------
    # Aggregates
    # --------------------------------------------------------------------
    RunSummary:
      type: object
      description: List-view shape (omits `dag` and per-job `steps`).
      required:
        - id
        - status
        - version
        - submitted_at
        - deck_jobs_summary
      additionalProperties: false
      properties:
        id:
          type: string
        status:
          $ref: '#/components/schemas/RunStatus'
        version:
          type: integer
          format: int64
        submitted_at:
          type: string
          format: date-time
        terminal_at:
          type: string
          format: date-time
          nullable: true
          description: |
            Stamped only when the run reaches a terminal status (`COMPLETED`
            or `CANCELLED`). Null while in `PENDING`, `RUNNING`, `FAILED`,
            or `AMBIGUOUS` — including FAILED runs that are awaiting an
            operator decision.
        deck_jobs_summary:
          $ref: '#/components/schemas/DeckJobsSummary'
    Run:
      type: object
      description: Run detail. Same shape returned by `POST /api/runs` and operator mutations.
      required:
        - id
        - status
        - version
        - submitted_at
        - dag
        - deck_jobs
      additionalProperties: false
      properties:
        id:
          type: string
        status:
          $ref: '#/components/schemas/RunStatus'
        version:
          type: integer
          format: int64
        submitted_at:
          type: string
          format: date-time
        terminal_at:
          type: string
          format: date-time
          nullable: true
          description: |
            Stamped only when the run reaches a terminal status (`COMPLETED`
            or `CANCELLED`). Null while in `PENDING`, `RUNNING`, `FAILED`,
            or `AMBIGUOUS` — including FAILED runs that are awaiting an
            operator decision.
        dag:
          $ref: '#/components/schemas/DagSubmission'
        deck_jobs:
          type: array
          items:
            $ref: '#/components/schemas/DeckJob'
...
    StateSnapshot:
      type: object
      description: |
        Response for `GET /api/state`. When `since_seq=0` (bootstrap)
        `runs` and `decks` are populated with the current fleet view;
        in steady state (`since_seq > 0`) `decks` is omitted and
        `decks_delta` carries only the touched-since slice. `server_seq`
        is the highest event seq known to the orchestrator at response
        time — clients use it as the next request's `since_seq`.
      required:
        - server_seq
        - events
      additionalProperties: false
      properties:
        server_seq:
          type: integer
          format: int64
        events:
          type: array
          items:
            $ref: '#/components/schemas/Event'
        runs:
          type: array
          description: Bootstrap-only. Omitted from delta responses.
          items:
            $ref: '#/components/schemas/RunSummary'
        decks:
          type: array
          description: |
            Bootstrap-only authoritative slice for every deck. Omitted
            from delta responses (since_seq > 0); use `decks_delta` there.
          items:
            $ref: '#/components/schemas/Deck'
        decks_delta:
          type: array
          description: |
            Delta-only. Decks whose row was touched since `since_seq`:
            union of (a) decks referenced by any event with seq >
            since_seq and (b) decks whose `last_heartbeat_at` advanced
            inside the recent freshness window (orchestrator-chosen,
            currently 5s). Clients merge by `id` into the existing
            decks cache instead of replacing wholesale. Bounds steady
            poll cost by activity rather than fleet size (see S3 in
            analysis/inefficiencies/inefficiencies.md).
          items:
            $ref: '#/components/schemas/Deck'
    RunStateSnapshot:
      type: object
      description: |
        Response for `GET /api/runs/{id}/state`. Same semantics as
        `StateSnapshot` but scoped to one run; `events` only includes
        rows whose `run_id` matches. `run` is populated on bootstrap.
      required:
        - server_seq
        - events
      additionalProperties: false
      properties:
        server_seq:
          type: integer
          format: int64
        events:
          type: array
          items:
            $ref: '#/components/schemas/Event'
        run:
          allOf:
            - $ref: '#/components/schemas/Run'
          nullable: true
          description: Bootstrap-only. Omitted from delta responses.

    # --------------------------------------------------------------------
    # Operator action requests
    # --------------------------------------------------------------------
...
    ErrorResponse:
      type: object
      description: Outer envelope wrapping every non-2xx response.
      required: [error]
      additionalProperties: false
      properties:
        error:
          $ref: '#/components/schemas/Error'